splits things into routers
