+++ README.md
@@ -3,3 +3,8 @@ How to startup mongodb
 $ pathtomongo/bin/mongod --dbpath=/pathtodbfolder/mongodb-data
 
  /Users/marlenebowles/mongodb
+
+ - figure out how to rename app
+ - figure out httpstatuses
+ - add groups that is related to users
+ - add users to tasks
